resolving cyclic dependencies
