+++ app/client/src/PluginActionEditor/components/PluginActionSettings/SettingsPopover.tsx
@@ -13,7 +13,7 @@ import {
   isPluginActionSettingsOpen,
   openPluginActionSettings,
 } from "../../store";
-import { THEME } from "../../constants/PluginActionConstants";
+import { THEME } from "../../types/PluginActionTypes";
 import { type DocsLink, openDoc } from "constants/DocumentationLinks";
 import { ToolbarSettingsPopover } from "IDE";
