+++ app/client/src/components/editorComponents/CodeEditor/EditorConfig.ts
@@ -1,8 +1,10 @@
 import type CodeMirror from "codemirror";
 import type { EntityTypeValue } from "ee/entities/DataTree/types";
-import type { DataTree } from "entities/DataTree/dataTreeTypes";
+import type {
+  DataTree,
+  EntityNavigationData,
+} from "entities/DataTree/dataTreeTypes";
 import type { AutocompleteDataType } from "utils/autocomplete/AutocompleteDataType";
-import type { EntityNavigationData } from "selectors/navigationSelectors";
 import type { ExpectedValueExample } from "utils/validation/common";
 
 import { editorSQLModes } from "./sql/config";
