+++ app/client/src/components/editorComponents/CodeEditor/MarkHelpers/entityMarker.ts
@@ -1,7 +1,7 @@
 import type {
   EntityNavigationData,
   NavigationData,
-} from "selectors/navigationSelectors";
+} from "entities/DataTree/dataTreeTypes";
 import type { MarkHelper } from "../EditorConfig";
 
 export const NAVIGATE_TO_ATTRIBUTE = "data-navigate-to";
