+++ app/client/src/constants/AppsmithActionConstants/formConfig/ApiSettingsConfig.ts
@@ -9,7 +9,7 @@ import {
 import {
   RUN_BEHAVIOR,
   RUN_BEHAVIOR_VALUES,
-} from "PluginActionEditor/constants/PluginActionConstants";
+} from "PluginActionEditor/types/PluginActionTypes";
 
 export default [
   {
