+++ app/client/src/entities/DataTree/dataTreeTypes.ts
@@ -3,6 +3,7 @@ import type {
   DataTreeEntityObject,
   UnEvalTreeEntityObject,
   AppsmithEntity,
+  EntityTypeValue,
 } from "ee/entities/DataTree/types";
 
 export interface unEvalAndConfigTree {
@@ -23,3 +24,23 @@ export type UnEvalTreeEntity = UnEvalTreeEntityObject | AppsmithEntity;
 export interface UnEvalTree {
   [entityName: string]: UnEvalTreeEntity;
 }
+
+export interface NavigationData {
+  name: string;
+  id: string;
+  type: EntityTypeValue;
+  isfunction?: boolean;
+  url: string | undefined;
+  navigable: boolean;
+  children: EntityNavigationData;
+  key?: string;
+  pluginName?: string;
+  pluginId?: string;
+  isMock?: boolean;
+  datasourceId?: string;
+  actionType?: string;
+  widgetType?: string;
+  value?: boolean | string;
+}
+
+export type EntityNavigationData = Record<string, NavigationData>;
