+++ app/client/src/pages/Editor/JSEditor/JSEditorToolbar/types.ts
@@ -1,6 +1,6 @@
 import type { JSAction } from "entities/JSCollection";
 import type { DropdownOption } from "@appsmith/ads-old";
-import type { ActionRunBehaviourType } from "PluginActionEditor/constants/PluginActionConstants";
+import type { ActionRunBehaviourType } from "PluginActionEditor/types/PluginActionTypes";
 
 export interface OnUpdateSettingsProps {
   value: ActionRunBehaviourType;
