+++ app/client/src/selectors/navigationSelectors.ts
@@ -1,4 +1,3 @@
-import type { EntityTypeValue } from "ee/entities/DataTree/types";
 import {
   ACTION_TYPE,
   JSACTION_TYPE,
@@ -34,26 +33,7 @@ import { PluginType } from "entities/Plugin";
 import type { StoredDatasource } from "entities/Action";
 import type { Datasource } from "entities/Datasource";
 import { getModuleInstanceNavigationData } from "ee/utils/moduleInstanceNavigationData";
-
-export interface NavigationData {
-  name: string;
-  id: string;
-  type: EntityTypeValue;
-  isfunction?: boolean;
-  url: string | undefined;
-  navigable: boolean;
-  children: EntityNavigationData;
-  key?: string;
-  pluginName?: string;
-  pluginId?: string;
-  isMock?: boolean;
-  datasourceId?: string;
-  actionType?: string;
-  widgetType?: string;
-  value?: boolean | string;
-}
-
-export type EntityNavigationData = Record<string, NavigationData>;
+import type { EntityNavigationData } from "entities/DataTree/dataTreeTypes";
 
 export const getModulesData = createSelector(
   getModuleInstances,
