+++ app/client/src/utils/NavigationSelector/WidgetChildren.ts
@@ -2,7 +2,7 @@ import type { WidgetEntity } from "ee/entities/DataTree/types";
 import type { DataTree } from "entities/DataTree/dataTreeTypes";
 import { ENTITY_TYPE } from "ee/entities/DataTree/types";
 import { builderURL } from "ee/RouteBuilder";
-import type { EntityNavigationData } from "selectors/navigationSelectors";
+import type { EntityNavigationData } from "entities/DataTree/dataTreeTypes";
 import { createNavData } from "./common";
 
 export const getWidgetChildrenNavData = (
